backend/requirements.txt
# Python dependencies
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0

# AI/ML
transformers==4.36.0
torch==2.1.0
textblob==0.17.1
scikit-learn==1.3.2
numpy==1.26.2

# Web Processing
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0
lxml==4.9.3

# Database
sqlalchemy==2.0.23
flask-sqlalchemy==3.0.5

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2

# Utilities
validators==0.22.0